pyproject.toml
[project]
name = "streamlit1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datetime>=5.5",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "streamlit>=1.50.0",
]
